+++ src/edit/material/interface.ts
@@ -1,0 +1,17 @@
+export enum MaterialType {
+    Default = 'DefaultPhong',
+    Basic = 'Basic',
+    Lambert = 'Lambert',
+    Blinn = 'Blinn',
+    Phong = 'Phong',
+    BlinnPhong = 'BlinnPhong',
+    Standard = 'Standard',      // PBR
+    Normal = 'Normal',
+    Depth = 'Depth',
+    Displacement = 'Displacement',
+    Shadow = 'Shadow',
+    Shader = 'Shader', // 自定义 shader 材质
+    Line = 'Line',
+    Points = 'Points',
+    Sprite = 'Sprite', 
+}